+++ src/writer.rs
@@ -67,6 +67,7 @@ impl<W: Write> EventWriter<W> {
             XmlEvent::CData(content) => self.emitter.emit_cdata(&mut self.sink, content),
             XmlEvent::Characters(content) => self.emitter.emit_characters(&mut self.sink, content),
             XmlEvent::RawCharacters(content) => self.emitter.emit_raw_characters(&mut self.sink, content),
+            XmlEvent::Doctype(content) => self.emitter.emit_raw_characters(&mut self.sink, content),
         }
     }