+++ Changelog.md
@@ -1,5 +1,6 @@
 ## Version 1.0.0
 
+* Added `Doctype` event
 * Marked structs as `#[non_exhaustive]`
 * Merged `ParserConfig2` back into `ParserConfig`
 * Added option to the writer to pass through XML markup unmodified